speech_text_tool/requirements.txt
flask
speechrecognition
pyttsx3
pytesseract
pillow
